requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
cryptography==41.0.4
PyJWT==2.8.0
gunicorn==21.2.0